feat(Medicines): Refactor medicine details management with new components and routing
- Rename MedicinesDetailPage to MedicinesAdminDetailPage for clarity.
- Introduce MedicinesAdminDetailContent, MedicinesAdminDetailSidebar, and MedicinesAdminDetailToolbar for better organization of medicine details.
- Update routing to reflect new structure and improve navigation for medicine details.
- Remove deprecated MedicinesDetailPage component and related files.
- Enhance API integration for fetching detailed medicine information.
