+++ src/page/admin/index.ts
@@ -1,4 +1,4 @@
 export { default as MedicinesAdminPage } from "./medicines";
-export { MedicinesDetailPage } from "./medicines";
+export { default as MedicinesAdminDetailPage } from "./medicines/[id]";
 export { default as UsersAdminPage } from "./users";
