+++ src/page/admin/medicines/[id]/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./page";
