+++ src/page/store/index.ts
@@ -1,5 +1,5 @@
 export { default as CategoriesPage } from "./categories/page";
-export { default as MedicineDetailsPage } from "./medicines/[id]/page";
+export { default as MedicinesDetailPage } from "./medicines/[id]/page";
 export { default as MedicinesPage } from "./medicines/page";
 export { default as OrdersPage } from "./orders/page";
 export { default as RootPage } from "./root/page";
