+++ src/page/store/medicines/index.ts
@@ -1,1 +1,2 @@
+export { default as MedicinesDetailPage } from "./[id]/page";
 export { default } from "./page";
